improving carousel in client agenda
