--- FrontEnd/src/app/src/components/announcement-carousel/announcement-carousel.component.html
@@ -1,10 +1,0 @@
-<app-carousel [visibleItems]="1">
-  @for (announcement of announcements ; track announcement.id){
-    <ng-template #carouselItem >
-      <mat-card (click)="open(announcement)">
-        <mat-card-title>{{ announcement.title }}</mat-card-title>
-        <mat-card-content>{{ announcement.body }}</mat-card-content>
-      </mat-card>
-    </ng-template>
-  }
-</app-carousel>
