--- FrontEnd/src/app/src/components/announcement-carousel/announcement-carousel.component.scss
@@ -1,35 +1,0 @@
-:host {
-  display: flex;
-  justify-content: center;
-  align-items: center;
-  height: 100%;
-  width: 100%;
-}
-
-app-carousel {
-  width: 100%;
-  height: 100%;
-  display: flex;
-  justify-content: center;
-  align-items: center;
-}
-
-
-mat-card {
-  width: 100%;
-  height: 100%;
-  display: flex;
-  flex-direction: column;
-  justify-content: start;
-  align-items: center;
-  border-radius: 10px;
-  margin: 0;
-  padding: 0 60px;
-  box-shadow: 0 4px 8px 0 rgba(0, 0, 0, 0.2);
-}
-
-mat-card-title {
-  font-size: 1.5em;
-  margin-top: 30px;
-  margin-bottom: 50px;
-}
