--- FrontEnd/src/app/src/components/carousel/carousel-item.directive.ts
@@ -1,9 +1,0 @@
-import { Directive, TemplateRef } from '@angular/core';
-
-@Directive({
-  standalone: true,
-  selector: '[carouselItem]'
-})
-export class CarouselItemDirective {
-  constructor(public templateRef: TemplateRef<any>) {}
-}
