--- FrontEnd/src/app/src/components/carousel/carousel.component.html
@@ -1,17 +1,0 @@
-<div class="carousel-container">
-  <div class="carousel-wrapper">
-    <button mat-icon-button class="carousel-button left" (click)="previous()">
-      <mat-icon>chevron_left</mat-icon>
-    </button>
-    <div class="carousel-content">
-      <ng-container *ngFor="let item of getVisibleItems(); let i = index">
-        <div [@carouselAnimation] class="carousel-item" [ngStyle]="{'flex': '0 0 calc(100% / ' + visibleItems + ')'}">
-          <ng-container *ngTemplateOutlet="item"></ng-container>
-        </div>
-      </ng-container>
-    </div>
-    <button mat-icon-button class="carousel-button right" (click)="next()">
-      <mat-icon>chevron_right</mat-icon>
-    </button>
-  </div>
-</div>
