+++ FrontEnd/src/app/src/components/material-carousel/material-carousel.component.scss
@@ -1,0 +1,99 @@
+@use "../../../../variables.scss";
+
+:host {
+  display: block;
+  width: 100%;
+  height: 100%;
+}
+
+.carousel-container {
+  position: relative;
+  display: flex;
+  align-items: center;
+  width: 100%;
+  height: 100%;
+  border-radius: 10px;
+
+  .custom-tab-group {
+    width: 100%;
+    height: 100%;
+    border-radius: 10px;
+    ::ng-deep mat-tab-header {
+      display: none !important; /* Oculta el encabezado */
+    }
+
+    ::ng-deep .mat-tab-group {
+      height: 100% !important;
+      border-radius: 10px;
+    }
+
+    ::ng-deep .mat-mdc-tab-body-wrapper {
+      height: 100% !important;
+      border-radius: 10px;
+    }
+
+    ::ng-deep .mat-tab-body {
+      height: 100% !important;
+      border-radius: 10px;
+    }
+  }
+
+  .slide-content {
+    display: flex;
+    flex-direction: column;
+    align-items: center;
+    justify-content: center;
+    height: 100%;
+    width: 100%;
+    box-shadow: 0 2px 4px rgba(0, 0, 0, 0.2);
+    border-radius: 10px;
+  }
+
+  .navigation-buttons-left,
+  .navigation-buttons-right {
+    position: absolute;
+    top: 50%;
+    display: flex;
+    justify-content: space-between;
+    align-items: center;
+    transform: translateY(-50%);
+    z-index: 1;
+  }
+
+  .navigation-buttons-left {
+    left: 0;
+    margin-left: 10px;
+  }
+
+  .navigation-buttons-right {
+    right: 0;
+    margin-right: 10px;
+  }
+
+  mat-card {
+    width: 100%;
+    height: 100%;
+    display: flex;
+    flex-direction: column;
+    justify-content: start;
+    align-items: center;
+    border-radius: 10px;
+    margin: 0;
+    padding: 0 60px;
+    box-shadow: 0 4px 8px 0 rgba(0, 0, 0, 0.2);
+    color: variables.$white-color;
+    border-radius: 10px;
+  }
+
+  mat-card-title {
+    font-size: 1.5em;
+    margin-top: 30px;
+    margin-bottom: 20px;
+    border-radius: 10px;
+  }
+
+  .card:hover {
+    cursor: pointer;
+    filter: brightness(1.1);
+  }
+}
