+++ FrontEnd/src/app/src/components/material-carousel/material-carousel.component.spec.ts
@@ -1,0 +1,23 @@
+import { ComponentFixture, TestBed } from '@angular/core/testing';
+
+import { MaterialCarouselComponent } from './material-carousel.component';
+
+describe('MaterialCarouselComponent', () => {
+  let component: MaterialCarouselComponent;
+  let fixture: ComponentFixture<MaterialCarouselComponent>;
+
+  beforeEach(async () => {
+    await TestBed.configureTestingModule({
+      imports: [MaterialCarouselComponent]
+    })
+    .compileComponents();
+
+    fixture = TestBed.createComponent(MaterialCarouselComponent);
+    component = fixture.componentInstance;
+    fixture.detectChanges();
+  });
+
+  it('should create', () => {
+    expect(component).toBeTruthy();
+  });
+});
