+++ FrontEnd/src/app/src/layout/client-agenda/client-agenda.component.html
@@ -40,7 +40,7 @@
             </mat-form-field>
           </div>
           <div class="quick-actions">
-            <app-announcement-carousel></app-announcement-carousel>
+            <app-material-carousel></app-material-carousel>
           </div>
         </section>
         <section class="botom-section">
